pyproject.toml
[project]
name = "kid-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
    "supabase>=2.15.1",
]
